requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
streamlit==1.30.0
pymongo[srv]==3.12.0
